This is balanced distributed bubble sort
First we take input from File input.txt.
Result goes to answeris.txt
Time taken goes to timeis.txt
There is a limit of 50000 elements for the array
First Process 0 (root) takes input from the input file. The input is in
form of #_of_elements and then the elements.
The number of elements is taken first and then broadcasted. Then the
actual array is broadcasted.
The process of choice is odd even transposition sort, as bubble sort is
inherently serial.
We divide the elements by number of processes. This may leave remainder,
and so we use MPI_Gatherv for the final gathering of all the parts.
Each process except the last is of the size as mentioned above. the last
process is given all the remaining elements.
Then for each local array "Serial Bubble Sort" is used. After this odd
even transposition is used.
For each phase, every process communicates its partener, and then merges
2 sorted array using 2 pointer process.
FInallly all the local arrays are sent to the global_arr_2 of process 0.
This is the result.
//Answer to part c
the theoretical speedup, assuming no overheads and one processor per
element, the time is(N is the number of elements)
If time taken by one processor is T(1) = O(N^2), then with increase in
processors, the time T(n) = (time_for_each_comparison)N + time_to_merge =
C * O(N) + O(N);
SO T(N) = O(N).
The speedup then is = O(N^2) / O(N), which for very large N =
(Some_Constant) * N
//Answer to part d
//Analysis of time taken
for # 1 threads, time is 0.000964
for # 2 threads, time is 0.000375
for # 3 threads, time is 0.002276
for # 4 threads, time is 0.013270
for # 5 threads, time is 0.002478
for # 6 threads, time is 0.001882
I am running Ubuntu VM on Windows OS with 2 cores. So the max speed is at
2 threads. Also there is extra overhead involved because of vms